Fixup Sandbox apps QtHelloWorld and QtSimpleAppTest to work with qt 15.5.0 and android

+++ SandboxApps/QtSimpleAppTest/QtSimpleTestApp/mainwindow.cpp
@@ -14,8 +14,8 @@ MainWindow::MainWindow(QWidget *parent) :
 
 
     char textBuf[16000];
-    sprintf( textBuf, "sizeof void * %d\n, int %d\n long %d\n longlong %d\n float %d\n double %d shared value %d static value %d\n",
-             sizeof(void *), sizeof(int), sizeof(long), sizeof(long long), sizeof(float), sizeof(double),
+    sprintf( textBuf, "sizeof void * %d\n, int %d\n long %d\n longlong %d\n float %d\n double %d shared value %zd static value %d\n",
+             (int)sizeof(void *), (int)sizeof(int), (int)sizeof(long), (int)sizeof(long long), (int)sizeof(float), (int)sizeof(double),
              staticLib.getStaticValue(),
              sharedLib.getSharedValue()
              );
